requirements.txt
flask==2.3.3
sentence-transformers==2.2.2
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
faiss-cpu==1.7.4